Use HoneySQL to generate INSERT SQL, instead of clojure.java.jdbc
Inserts were being done with jdbc/insert-multi!, which internally generates the
necessary SQL statements. This means HoneySQL is bypassed for inserts, and so
HoneySQL features like calling SQL functions with (sql/call ,,,) are unavailable.

This patch instead makes HoneySQL generate the INSERT statements, and executes
them with the lower level jdbc/db-do-prepared-return-keys. When inserting
multiple rows at once it executes separate INSERT statements, rather than a
single statement with multiple VALUES. This behavior is the same with
jdbc/insert-multi!. It's otherwise not possible to retrieve the primary keys of
the inserted rows. (This is a limitation of jdbc/db-do-prepared-return-keys, not
of JDBC itself).

I've added test-cases for db/insert-many! and db/simple-insert-many!, since
those were still missing, as well as an extra test case for db/insert! which
demonstrates the use of an SQL function call.