Make defmodel take protocol implementations
Pass implementations of protocols directly to `defmodel`, the same way you would
do in `defrecord` or `deftype`. (fixes #9)

Also provides implementations for `empty` (fixes #1) and `apply` (fixes #2).

This updates the documentation, promoting this way of doing things as the
preferred method (vs using extend...merge...IModelDefaults), but still leaves in
an example of how to do things the old-fashioned way.

The defining-models doc has been expanded slightly to further clarify the
relationship between `User` and `(User 1`).

The fill column in that document was irregular but seemed to hover around 120,
so I went with breaking lines at 120 characters. It might be best to ignore
whitespace changes when reviewing this.

Note that as it stands you can implement protocols in `defrecord`, but not
interfaces, since the forms are inserted in `extend` rather than in `defrecord`
itself. To fix this we'll have to separate extensions to IModel from other
interface/protocol implementations.