Remove classpath scanning in automagic batched hydration
This code is pretty slow with large classpaths, taking over 10 seconds
in Metabase. This code is also broken on JDK 9 due to the classloader
changes, `classpath/classpath` will always return the empty
list. With this removed, all of the Metabase tests still pass and we
can save that 10 second load time the first time hydration is called.